# sql: improve inverted index validation performance