@vivekmenezes 2019-03-14
@lucy-zhang it's probably worth adding a benchmark so that future changes don't cause a regression in performance. Thanks!