@lucy-zhang 2019-05-13
I added a benchmark for a JSON object containing mostly scalars and few arrays:
```
name                           old time/op  new time/op  delta
JSONNumInvertedIndexEntries-8  6.93µs ± 0%  2.92µs ± 1%  -57.91%  (p=0.008 n=5+5)
```